+++ src/Controller/Api/CheckFilerobotUuidExistApiController.php
@@ -12,7 +12,7 @@ use Shopware\Core\Framework\Routing\Annotation\RouteScope;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @RouteScope(scopes={"api"})
+ * @Route(defaults={"_routeScope"={"api"}})
  */
 class CheckFilerobotUuidExistApiController extends AbstractController
 {