+++ CHANGELOG.md
@@ -15,4 +15,4 @@
 - Update Filerobot Widget library
 
 # 1.0.6
-- Update Filerobot Widget library
+- Update Filerobot Widget library and fix bug